Clean up code action code
- Move code action controller to own class
- Remove unused properties
- Use flatMap
